+++ resources/views/components/chart/legend.blade.php
@@ -1,8 +1,13 @@
 @props([
     'position' => 'top',   // top | bottom
+    'show' => true,
 ])
 @php
-    app(\KoreUi\Charts\ChartContext::class)->current('legend')->legend = [
-        'position' => in_array($position, ['top', 'bottom'], true) ? $position : 'top',
-    ];
+    $frame = app(\KoreUi\Charts\ChartContext::class)->current('legend');
+
+    if (filter_var($show, FILTER_VALIDATE_BOOL)) {
+        $frame->legend = [
+            'position' => in_array($position, ['top', 'bottom'], true) ? $position : 'top',
+        ];
+    }
 @endphp